+++ src/app/pages/core/agreement/update/document/document.component.ts
@@ -157,12 +157,28 @@ export class DocumentComponent implements OnInit {
   }
 
   removeFile(index: number, item: FileModel) {
-    this.filesHttpService.remove(item.id!, true,this.formInput.id!).subscribe(response => {
-      this.formInput.enablingDocuments.splice(index, 1);
-      this.types.push(item.type!);
 
-      this.form.patchValue(this.formInput.enablingDocuments);
+    this.confirmationService.confirm({
+      key: 'confirmDialog',
+      message: '¿Está seguro de eliminar de manera definitiva el archivo?',
+      header: '',
+      icon: PrimeIcons.TRASH,
+      acceptLabel: "Si",
+      rejectLabel: "No",
+      rejectButtonStyleClass: "p-button-text",
+      accept: () => {
+        this.filesHttpService.remove(item.id!, true,this.formInput.id!).subscribe(response => {
+          this.formInput.enablingDocuments.splice(index, 1);
+          this.types.push(item.type!);
+
+          this.form.patchValue(this.formInput.enablingDocuments);
+        });
+      }
     });
+  }
+
+  download(file: FileModel) {
+    this.filesHttpService.downloadFile(file);
   }
 
   get typeField(): AbstractControl {
